Major performance improvement, and logic change
I am now using a dynamic time-step to manage both the movement and updating of the snake, which means that the large delay placed within the application is now gone. with this fix, i can now have the snake update with only one extra body part spawned, instead of the 6 that were spawned previously. this means that the performance is improved overall with this change as well.